@@ -1,48 +1,42 @@
 {
   "name": "nuxt-hk-weather",
   "version": "1.0.0",
+  "type": "module",
   "private": true,
   "engines": {
     "node": "18.x"
   },
   "scripts": {
-    "dev": "nuxt",
+    "dev": "nuxt dev",
     "build": "nuxt build",
     "start": "nuxt start",
     "export": "nuxt export",
     "serve": "nuxt serve",
-    "lint:js": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "lint:style": "stylelint **/*.{vue,css} --ignore-path .gitignore",
+    "lint:js": "eslint --ext .js,.vue .",
+    "lint:style": "stylelint **/*.{vue,css}",
     "lint": "npm run lint:js && npm run lint:style",
-    "test:e2e": "wdio wdio.conf.js",
     "test:cypress": "npx cypress run",
     "cypress:open": "npx cypress open"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.13.6",
-    "@nuxtjs/i18n": "^7.3.1",
-    "@nuxtjs/pwa": "^3.3.5",
-    "@nuxtjs/sentry": "^8.0.7",
-    "date-fns": "^2.30.0",
-    "node-fetch-native": "^1.6.4",
-    "nuxt": "^2.18.1",
+    "@nuxt/eslint": "^1.2.0",
+    "@nuxtjs/i18n": "^9.3.1",
+    "date-fns": "^4.1.0",
+    "nuxt": "^3.16.0",
     "tachyons": "^4.12.0"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.19.1",
-    "@nuxtjs/eslint-config": "^8.0.0",
-    "@nuxtjs/eslint-module": "^3.1.0",
-    "@nuxtjs/stylelint-module": "^4.1.0",
-    "cypress": "^13.11.0",
-    "eslint": "^8.3.0",
+    "@nuxtjs/stylelint-module": "^5.2.0",
+    "cypress": "^14.2.0",
+    "eslint": "^9.22.0",
     "eslint-config-prettier": "^8.6.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-nuxt": "^2.0.0",
+    "eslint-plugin-cypress": "^4.2.0",
     "eslint-plugin-prettier": "^3.4.1",
     "postcss-html": "^1.5.0",
     "prettier": "^2.8.3",
-    "stylelint": "^14.16.1",
-    "stylelint-config-prettier": "^9.0.4",
-    "stylelint-config-standard": "^29.0.0"
+    "stylelint": "^16.16.0",
+    "stylelint-config-standard": "^37.0.0",
+    "stylelint-config-standard-scss": "^14.0.0",
+    "typescript": "^5.8.2"
   }
 }
