@@ -1,5 +1,6 @@
-module.exports = {
-  extends: ['stylelint-config-standard', 'stylelint-config-prettier'],
+/** @type {import('stylelint').Config} */
+export default {
+  extends: ['stylelint-config-standard'],
   // add your custom config here
   // https://stylelint.io/user-guide/configuration
   rules: {},
