@@ -1,3 +1,16 @@
+<script setup>
+defineProps({
+  id: {
+    type: String,
+    default: 'element-districts',
+  },
+  districts: {
+    type: Array,
+    default: () => [],
+  },
+})
+</script>
+
 <template>
   <div :id="id">
     <div
@@ -14,18 +27,3 @@
     </div>
   </div>
 </template>
-
-<script>
-export default {
-  props: {
-    id: {
-      type: String,
-      default: 'element-districts',
-    },
-    districts: {
-      type: Array,
-      default: () => [],
-    },
-  },
-}
-</script>
